requirements.txt
pyyaml
pytest
pytest-selenium
allure-pytest
pytest-xdist
pytest-failed-screenshot
Faker